Rename utils/registry_discovery to utils/registry_jwks
The package name "registry_discovery" was confusing because it evokes
federation discovery, which is unrelated. The package's actual purpose
is looking up issuer URLs and JWKS for namespaces registered at the
Pelican registry. Rename the directory and package to "registry_jwks"
to make this intent explicit.
