+++ cmd/token.go
@@ -41,7 +41,7 @@ import (
 	"github.com/pelicanplatform/pelican/server_structs"
 	"github.com/pelicanplatform/pelican/token"
 	"github.com/pelicanplatform/pelican/token_scopes"
-	"github.com/pelicanplatform/pelican/utils/registry_discovery"
+	"github.com/pelicanplatform/pelican/utils/registry_jwks"
 )
 
 var (
@@ -132,7 +132,7 @@ func splitClaim(claim string) (string, string, error) {
 
 // Given some issuer and a set of KIDs, determine whether the issuer's JWKS contains a key with a matching KID
 func issuerMatchesKey(issuer string, kidSet map[string]struct{}) (bool, error) {
-	remoteJWKS, err := registry_discovery.GetJWKSFromIssUrl(issuer)
+	remoteJWKS, err := registry_jwks.GetJWKSFromIssUrl(issuer)
 	if err != nil {
 		return false, err
 	}
@@ -159,7 +159,7 @@ func getIssuer(directorInfo server_structs.DirectorResponse, kidSet map[string]s
 
 	// Comb through the JWKS from each issuer to find which matches the signing key
 	for _, issuer := range directorInfo.XPelAuthHdr.Issuers {
-		remoteJWKS, err := registry_discovery.GetJWKSFromIssUrl(issuer.String())
+		remoteJWKS, err := registry_jwks.GetJWKSFromIssUrl(issuer.String())
 		if err != nil {
 			log.Warningf("Unable to get JWKS from issuer URL %s: %v; skipping", issuer, err)
 			continue
