+++ lotman/lotman_ui.go
@@ -40,7 +40,7 @@ import (
 	"github.com/pelicanplatform/pelican/token"
 	"github.com/pelicanplatform/pelican/token_scopes"
 	"github.com/pelicanplatform/pelican/utils"
-	"github.com/pelicanplatform/pelican/utils/registry_discovery"
+	"github.com/pelicanplatform/pelican/utils/registry_jwks"
 )
 
 type LotAction string
@@ -56,7 +56,7 @@ func tokenSignedByAuthorizedCaller(strToken string, authorizedCallers *[]string)
 	ownerFound := false
 	var tok jwt.Token
 	for _, owner := range *authorizedCallers {
-		kSet, err := registry_discovery.GetJWKSFromIssUrl(owner)
+		kSet, err := registry_jwks.GetJWKSFromIssUrl(owner)
 		if err != nil {
 			log.Debugf("Error getting JWKS for owner %s: %v", owner, err)
 			continue
@@ -121,7 +121,7 @@ func VerifyNewLotToken(lot *Lot, strToken string) (bool, error) {
 			return false, err
 		}
 
-		kSet, err := registry_discovery.GetJWKSFromIssUrl(issuerUrl)
+		kSet, err := registry_jwks.GetJWKSFromIssUrl(issuerUrl)
 		if err != nil {
 			return false, errors.Wrap(err, "Error getting JWKS from issuer URL")
 		}
@@ -149,7 +149,7 @@ func VerifyNewLotToken(lot *Lot, strToken string) (bool, error) {
 
 		ownerFound := false
 		for owner := range ownersSet {
-			kSet, err := registry_discovery.GetJWKSFromIssUrl(owner)
+			kSet, err := registry_jwks.GetJWKSFromIssUrl(owner)
 
 			// Print the kSet as a string for debugging
 			kSetStr, _ := json.Marshal(kSet)
@@ -250,7 +250,7 @@ func VerifyNewLotToken(lot *Lot, strToken string) (bool, error) {
 	namespace := xPelicanNamespaceMap["namespace"]
 
 	// Get the issuer URL for that namespace
-	nsIssuerUrl, err := registry_discovery.GetNSIssuerURL(namespace)
+	nsIssuerUrl, err := registry_jwks.GetNSIssuerURL(namespace)
 	if err != nil {
 		return false, errors.Wrapf(err, errMsgPrefix+"no issuer could be found for namespace %s", namespace)
 	}
